+++ src/relayFeeCalculator/chain-queries/index.ts
@@ -1,2 +1,3 @@
 export * from "./baseQuery";
 export * from "./polygon";
+export * from "./factory";
